requirements.txt
ultralytics>=8.3.0
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.10.0
streamlit>=1.39.0
pandas>=2.2.2
matplotlib>=3.9.0
pyyaml>=6.0.1
tqdm>=4.66.0
onnx>=1.16.0
onnxruntime>=1.18.0

